# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NiChart API",
    "summary": "Backend API for the NiChart medical-imaging pipeline platform.",
    "description": "Provides endpoints for managing projects, uploading imaging data, running containerised processing pipelines (locally via Docker or on AWS Batch), and retrieving results. \n\n**Authentication**: In cloud mode every request requires a Cognito ID token in the ``Authorization: Bearer`` header. In local mode (``NICHART_EXECUTION_MODE=local``) authentication is bypassed.",
    "contact": {
      "name": "CBICA",
      "url": "https://github.com/CBICA/NiChart_Project"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects/{project_id}/jobs/pipelines": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Submit a pipeline run",
        "description": "Starts an asynchronous pipeline run for the given project. The server spawns a background task that executes each pipeline step in order, polling the job backend between steps. Use the returned ``run_id`` to track progress via ``GET /jobs/pipelines/{run_id}``.",
        "operationId": "submit_pipeline_projects__project_id__jobs_pipelines_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineRunSubmit"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunDetail"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid pipeline ID or parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Project or pipeline not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/pipelines": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List pipeline runs",
        "description": "Returns the authenticated user's pipeline runs in reverse-chronological order. Pass ``project_id`` to filter to a single project.",
        "operationId": "list_pipeline_runs_jobs_pipelines_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter runs to this project.",
              "title": "Project Id"
            },
            "description": "Filter runs to this project."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Maximum number of runs to return.",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum number of runs to return."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PipelineRunSummary"
                  },
                  "title": "Response List Pipeline Runs Jobs Pipelines Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/pipelines/{run_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get pipeline run detail",
        "description": "Returns the full run record including per-step status, timestamps, and any error messages. In cloud mode, while a step is actively waiting on the Batch queue, ``jobs_ahead`` and ``estimated_wait_seconds`` are populated by querying all jobs submitted to the queue before this one. These fields are null in local mode or once a job has left the queue.",
        "operationId": "get_pipeline_run_jobs_pipelines__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunDetail"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancel a pipeline run",
        "description": "Requests cancellation of a running pipeline. The currently executing tool step is cancelled on the backend; subsequent steps will not be submitted. No-op if the run is already in a terminal state.",
        "operationId": "cancel_pipeline_run_jobs_pipelines__run_id__delete",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/pipelines/{run_id}/logs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get pipeline run logs",
        "description": "Returns concatenated log output from all steps completed so far. Poll this endpoint alongside the status endpoint during a run.",
        "operationId": "get_pipeline_logs_jobs_pipelines__run_id__logs_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunLogs"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/pipelines/{run_id}/steps/{step_id}/chunks/{chunk_idx}/logs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get logs for a single parallel chunk",
        "description": "Returns the captured log output for one chunk of a parallelised pipeline step. Only available after the chunk has started executing. Use ``GET /jobs/pipelines/{run_id}/logs`` to retrieve all chunks concatenated.",
        "operationId": "get_chunk_logs_jobs_pipelines__run_id__steps__step_id__chunks__chunk_idx__logs_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Step Id"
            }
          },
          {
            "name": "chunk_idx",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Chunk Idx"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Get Chunk Logs Jobs Pipelines  Run Id  Steps  Step Id  Chunks  Chunk Idx  Logs Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "400": {
            "description": "Step was not run in parallel chunks.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Run, step, or chunk not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable description of the error."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "description": "Standard error response body."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PipelineRunDetail": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "description": "Unique run identifier (UUID)."
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "pipeline_id": {
            "type": "string",
            "title": "Pipeline Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "description": "Overall pipeline status."
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "current_step": {
            "type": "integer",
            "title": "Current Step",
            "description": "Index of the step currently executing (0-based).",
            "default": 0
          },
          "total_steps": {
            "type": "integer",
            "title": "Total Steps",
            "default": 0
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepStatus"
            },
            "type": "array",
            "title": "Steps"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Top-level error message if the run failed."
          },
          "jobs_ahead": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobs Ahead",
            "description": "Number of Batch jobs submitted before this one that are still waiting for compute capacity. Cloud mode only; null in local mode or when no step is currently pending on the Batch queue."
          },
          "estimated_wait_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Seconds",
            "description": "Estimated seconds until this job reaches the front of the queue, computed as the sum of (num_subjects \u00d7 time_per_subject_seconds) for each job submitted ahead of this one. Null when timing data is unavailable for any ahead job."
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "project_id",
          "pipeline_id",
          "status",
          "submitted_at"
        ],
        "title": "PipelineRunDetail",
        "description": "Full pipeline run record including per-step breakdown."
      },
      "PipelineRunLogs": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "logs": {
            "type": "string",
            "title": "Logs",
            "description": "Concatenated log output from all steps executed so far."
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "logs"
        ],
        "title": "PipelineRunLogs",
        "description": "Aggregated logs for all completed steps in a pipeline run."
      },
      "PipelineRunSubmit": {
        "properties": {
          "pipeline_id": {
            "type": "string",
            "title": "Pipeline Id",
            "description": "Pipeline identifier (matches a YAML basename in resources/pipelines/)."
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "description": "Parameter overrides applied to every step that accepts them."
          },
          "reuse_cached_steps": {
            "type": "boolean",
            "title": "Reuse Cached Steps",
            "description": "When True, steps whose inputs haven't changed since the last successful run are skipped. Set to False to force a full re-run.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "pipeline_id"
        ],
        "title": "PipelineRunSubmit",
        "description": "Request body to submit a pipeline run."
      },
      "PipelineRunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "description": "Unique run identifier (UUID)."
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "pipeline_id": {
            "type": "string",
            "title": "Pipeline Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "description": "Overall pipeline status."
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "current_step": {
            "type": "integer",
            "title": "Current Step",
            "description": "Index of the step currently executing (0-based).",
            "default": 0
          },
          "total_steps": {
            "type": "integer",
            "title": "Total Steps",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "project_id",
          "pipeline_id",
          "status",
          "submitted_at"
        ],
        "title": "PipelineRunSummary",
        "description": "Abbreviated pipeline run record for list responses."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}